+++ Source/Projects/EC/App/Regression/Func/VladERCA.cs
@@ -30,7 +30,7 @@ namespace BraneCloud.Evolution.EC.App.Regression.Func
         }
 
         public override string ToStringForHumans()
-        { return "n^" + (float)value; }
+        { return "n^" + value; }
 
         public override void Eval(IEvolutionState state,
             int thread,
